# local-hobbies-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Local Hobbies API
  description: API for a user-matching application based on hobbies, location, and availability.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://api.example.com/api
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /users/onboarding:
    get:
      tags:
        - Users
      summary: Get onboarding completion status
      description: Retrieve which onboarding steps must be performed
      responses:
        '200':
          description: Onboarding status retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserOnboardingIncompleteSection'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Users
      summary: Complete onboarding
      description: Submit fields that are required to complete onboarding and receive JWT token with user information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserOnboardingRequest'
      responses:
        '200':
          description: Onboarding completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users:
    get:
      tags:
        - Users
      summary: Get current user's account information
      description: Retrieve all account details for current user
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Users
      summary: Update current user information
      description: Update user's account details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User information updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Users
      summary: Delete current user's account
      description: Permanently delete the current user's account
      responses:
        '204':
          description: Account deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/profile-photo/upload-url:
    post:
      tags:
        - Users
      summary: Request profile photo upload URL
      description: Request a presigned URL to upload a profile photo to the cloud
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned URL received successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/homepage:
    get:
      tags:
        - Users
      summary: Get current user's homepage information
      description: Get summary info about the user's hobbies, schedule, and matches
      responses:
        '200':
          description: Homepage information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserHomepageResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/profile:
    get:
      tags:
        - Users
      summary: Get current user's profile
      description: Get current user's profile including public account information, hobbies, and photos. Fields that are set to private are returned as null.
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserProfileResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{otherUserId}/profile:
    get:
      tags:
        - Users
      summary: Get another user's public profile
      description: Get another user's profile including public account information, hobbies, and photos, and matching information. Fields that are set to private are returned as null.
      parameters:
        - name: otherUserId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtherUserProfileResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== Hobbies ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login, /auth/signup, or /users/onboarding
  schemas:
...
    AuthResponse:
      type: object
      required:
        - auth
        - user
      description: response to successful signup, login, and onboarding
      properties:
        auth:
          type: object
          required:
            - accesToken
            - tokenType
            - expirationTime
            - refreshToken
          properties:
            accessToken:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            tokenType:
              type: string
              example: Bearer
            expirationTime:
              type: string
              format: date-time
              description: ISO 8601 UTC timestamp
              example: "2026-02-08T12:00:00Z"
            refreshToken:
              type: string
              example: def502003a4b8c9d7e6f5a4b3c2d1e0f...
        user:
          type: object
          required:
            - id
            - onboardingComplete
          properties:
            id:
              type: integer
              example: 123
            onboardingComplete:
              type: boolean
              description: used to route to either homepage or onboarding page
              example: false

    # ==================== User, Homepage, & Profile Schemas ====================
    UserOnboardingIncompleteSection:
      type: object
      description: section of a user's account information that is required and incomplete
      required:
        - name
        - reason
      properties:
        name:
          type: string
          enum: [name, birthDate, genderMatched, location, publicContactInfo, hobbies, availabilities]
          example: hobbies
        reason:
          type: string
          enum: [noValue, minCountNotMet]
          example: minCountNotMet

#    UserOnboardingStatusResponse:
#      type: object
#      required:
#        - incompleteSections
#      description: list of sections that are yet to be completed for onboarding
#      properties:
#        incompleteSections:
#          type: array
#          items:
#            $ref: '#/components/schemas/UserOnboardingIncompleteSection'
    UserOnboardingRequest:
      type: object
      description: request to complete onboarding; contains only fields that are being updated
      properties:
        name:
          type: string
          example: Alice Smith
        birthDate:
          type: string
          format: date
          example: "1995-05-15"
        genderMatched:
          type: string
          enum: [Woman, Man, Nonbinary]
          example: Woman
        location:
          $ref: '#/components/schemas/GeoJsonPoint'
        publicContactInfo:
          type: string
          example: "@asmith123 on Instagram"
        hobbies:
          type: array
          items:
            $ref: '#/components/schemas/HobbyCreationRequest'
        availabilities:
          $ref: '#/components/schemas/AvailabilityOnboardingRequests'
    UserUpdateRequest:
      type: object
      description: contains only fields that are being changed
      properties:
        email:
          type: string
          example: alice@example.com
        name:
          type: string
          example: Alice Smith
        birthDate:
          type: string
          format: date
          example: "1995-05-15"
        genderDisplayed:
          type: string
          nullable: true
          description: null if being removed from account
          example: Woman
        genderMatched:
          type: string
          enum: [Woman, Man, Nonbinary]
          example: Woman
        bio:
          type: string
          nullable: true
          description: null if being removed from account
          example: Love hiking and coffee!
        location:
          $ref: '#/components/schemas/GeoJsonPoint'
        publicContactInfo:
          type: string
          example: "@asmith123 on Instagram"
        profilePhotoUrl:
          type: string
          description: omitted if not changed or if update was attempted by user but photo upload failed; null if being removed from account
          nullable: true
          example: https://example.com/photos/123.jpg
        showAge:
          type: boolean
          example: true
        showGenderDisplayed:
          type: boolean
          example: true
    UserResponse:
      type: object
      required:
        - username
        - email
        - name
        - birthDate
        - bio
        - location
        - publicContactInfo
        - profilePhotoUrl
        - genderMatched
        - showAge
        - showGenderDisplayed
      properties:
        username:
          type: string
          example: alice_smith
        email:
          type: string
          example: alice@example.com
        name:
          type: string
          example: Alice Smith
        birthDate:
          type: string
          format: date
          example: "1995-05-15"
        genderDisplayed:
          type: string
          nullable: true
          example: Woman
        bio:
          type: string
          nullable: true
          example: Love hiking and coffee!
        location:
          $ref: '#/components/schemas/GeoJsonPoint'
        publicContactInfo:
          type: string
          example: "@asmith123 on Instagram"
        profilePhotoUrl:
          type: string
          nullable: true
          example: https://example.com/photos/123.jpg
        genderMatched:
          type: string
          enum: [Woman, Man, Nonbinary]
          example: Woman
        showAge:
          type: boolean
          example: true
        showGenderDisplayed:
          type: boolean
          example: true
    UserHomepageResponse:
      type: object
      required:
        - user
        - hobbySummary
        - availabilitySummary
        - matchSummary
      properties:
        user:
          type: object
          required:
            - id
            - name
            - profilePhotoUrl
          properties:
            id:
              type: integer
              example: 123
            name:
              type: string
              example: Alice Smith
            profilePhotoUrl:
              type: string
              nullable: true
              example: https://example.com/photos/123.jpg
        hobbySummary:
          type: object
          required:
            - count
          properties:
            count:
              type: integer
              example: 5
        availabilitySummary:
          type: object
          required:
            - nextAvailability
          properties:
            nextAvailability: #change to booked meetup once implemented
              $ref: '#/components/schemas/OneTimeAvailabilityResponse'
        matchSummary:
          type: object
          required:
            - count
          properties:
            count:
              type: integer
              example: 13
    CurrentUserProfileResponse:
      type: object
      required:
        - name
        - age
        - genderDisplayed
        - bio
        - locationApproximate
        - publicContactInfo
        - profilePhotoUrl
        - hobbies
        - hobbyPhotos
      properties:
        name:
          type: string
          example: Alice Smith
        age:
          type: integer
          nullable: true
          description: null if showAge is false
          example: 29
        genderDisplayed:
          type: string
          nullable: true
          description: null if not set or if showGenderDisplayed is false
          example: Woman
        bio:
          type: string
          nullable: true
          example: Love hiking and coffee!
        locationApproximate:
          type: string
          example: San Francisco, CA
        publicContactInfo:
          type: string
          example: "@asmith123 on Instagram"
        profilePhotoUrl:
          type: string
          nullable: true
          example: https://example.com/photos/123.jpg
        hobbies:
          type: array
          items:
            $ref: '#/components/schemas/HobbyResponse'
        hobbyPhotos:
          type: array
          items:
            $ref: '#/components/schemas/HobbyPhotoResponse'
    OtherUserProfileResponse:
      type: object
      required:
        - id
        - name
        - age
        - genderDisplayed
        - bio
        - locationApproximate
        - publicContactInfo
        - profilePhotoUrl
        - lastSessionTime
        - hobbies
        - hobbyPhotos
        - isSavedMatch
        - overlappingHobbies
        - overlappingAvailabilities
      properties:
        id:
          type: integer
          example: 123
        name:
          type: string
          example: Alice Smith
        age:
          type: integer
          nullable: true
          description: null if showAge is false
          example: 29
        genderDisplayed:
          type: string
          nullable: true
          description: null if not set or if showGenderDisplayed is false
          example: Woman
        bio:
          type: string
          nullable: true
          example: Love hiking and coffee!
        locationApproximate:
          type: string
          example: San Francisco, CA
        publicContactInfo:
          type: string
          example: "@asmith123 on Instagram"
        profilePhotoUrl:
          type: string
          example: https://example.com/photos/123/profile.jpg
        lastSessionTime:
          type: string
          format: date-time
          description: ISO 8601 UTC timestamp
          example: "2026-02-08T12:00:00Z"
        hobbies:
          type: array
          items:
            $ref: '#/components/schemas/HobbyResponse'
        hobbyPhotos:
          type: array
          items:
            $ref: '#/components/schemas/HobbyPhotoResponse'
        isSavedMatch:
          type: boolean
          example: true
        overlappingHobbies:
          type: array
          items:
            $ref: '#/components/schemas/HobbyOverlapResponse'
        overlappingAvailabilities:
          type: array
          items:
            $ref: '#/components/schemas/AvailabilityOverlapResponse'

    # ==================== Hobby Schemas ====================
    HobbyCreationRequest:
      type: object
      required:
        - name
        - experienceLevel
      properties:
        name:
          type: string
          example: Hiking
        experienceLevel:
          type: string
          enum: [Beginner, Intermediate, Advanced]
          example: Intermediate
...
    HobbyResponse:
      type: object
      required:
        - id
        - name
        - category
        - experienceLevel
      properties:
        id:
          type: integer
          example: 49
        name:
          type: string
          example: Hiking
        category:
          type: string
          example: Fitness
        experienceLevel:
          type: string
          enum: [Beginner, Intermediate, Advanced]
          example: Intermediate
...
    HobbyPhotoResponse:
      type: object
      required:
        - id
        - hobbyId
        - photoUrl
        - caption
      properties:
        id:
          type: integer
          example: 335
        hobbyId:
          type: integer
          example: 49
        photoUrl:
          type: string
          example: https://example.com/photos/123/hobby135.jpg
        caption:
          type: string
          nullable: true
          example: Had a great time!
...
    HobbyOverlapResponse:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Hiking

    # ==================== Availability Schemas ====================
    OneTimeAvailabilityCreationRequest:
      type: object
      required:
        - location
        - start
        - duration
      properties:
        location:
          $ref: '#/components/schemas/GeoJsonPoint'
        startDate:
          type: string
          format: date
          example: "2026-03-15"
        start:
          type: string
          format: date-time
          description: ISO 8601 UTC timestamp
          example: "2026-02-08T12:00:00Z"
        duration:
          type: string
          description: ISO 8601 duration
          example: PT3H
...
    OneTimeAvailabilityResponse:
      type: object
      required:
        - id
        - userId
        - location
        - start
        - duration
      properties:
        id:
          type: integer
          example: 41
        userId:
          type: integer
          example: 123
        location:
          $ref: '#/components/schemas/GeoJsonPoint'
        start:
          type: string
          format: date-time
          description: ISO 8601 UTC timestamp
          example: "2026-02-08T12:00:00Z"
        duration:
          type: string
          description: ISO 8601 duration
          example: PT3H
    RecurringAvailabilityCreationRequest:
      type: object
      required:
        - location
        - ruleStart
        - frequency
        - startTime
        - duration
      properties:
        location:
          $ref: '#/components/schemas/GeoJsonPoint'
        ruleStart:
          type: string
          format: date
          example: "2026-01-01"
        ruleEnd:
          type: string
          format: date
          description: omitted if rule continues indefinitely
          example: "2026-12-31"
        frequency:
          type: string
          enum: [Weekly, Every two weeks, Monthly]
          example: Weekly
        startDayOfWeek:
          type: integer
          minimum: 1
          maximum: 7
          description: ISO 8601 day of week (1=Monday, 7=Sunday; for weekly/every two weeks); omitted if startDayOfMonth is submitted
        startDayOfMonth:
          type: integer
          minimum: 1
          maximum: 31
          description: day of month (for monthly); omitted if startDayOfWeek is submitted
        startTime:
          type: string
          format: time
          example: "14:00"
        duration:
          type: string
          description: ISO 8601 duration
          example: PT2H
...
    AvailabilityExceptionOnboardingCreationRequest:
      type: object
      description: request used during onboarding; omits recurringAvailabilityId, instead relying on JSON nesting to associate with recurring availability
      required:
        - exceptionDate
        - exceptionReason
        - isCancelled
      properties:
        exceptionDate:
          type: string
          format: date
          example: "2026-03-04"
        exceptionReason:
          type: string
          example: Traveling
        isCancelled:
          type: boolean
        overrideLocation:
          description: omitted if isCancelled = true
          allOf:
            - $ref: '#/components/schemas/GeoJsonPoint'
        overrideStartTime:
          type: string
          format: time
          description: omitted if isCancelled = true
        overrideDuration:
          type: string
          description: ISO 8601 duration; omitted if isCancelled = true
...
    AvailabilityOnboardingRequests:
      type: object
      description: availability creation requests submitted during onboarding
      required:
        - oneTimes
        - recurringsWithExceptions
      properties:
        oneTimes:
          type: array
          items:
            $ref: '#/components/schemas/OneTimeAvailabilityCreationRequest'
        recurringsWithExceptions:
          type: array
          items:
            $ref: '#/components/schemas/RecurringAvailabilityWithExceptions'
    RecurringAvailabilityWithExceptions:
      type: object
      description: recurring availability and its associated exceptions, submitted during onboarding
      required:
        - recurring
        - exceptions
      properties:
        recurring:
          $ref: '#/components/schemas/RecurringAvailabilityCreationRequest'
        exceptions:
          type: array
          items:
            $ref: '#/components/schemas/AvailabilityExceptionOnboardingCreationRequest'
...
    AvailabilityOverlapResponse:
      type: object
      required:
        - distanceKilometers
        - start
        - end
      properties:
        distanceKilometers:
          type: number
          format: double
          description: distance between the two availabilities
          example: 4.3
        start:
          type: string
          format: date-time
          description: ISO 8601 UTC timestamp
          example: "2026-02-08T12:00:00Z"
        end:
          type: string
          format: date-time
          description: ISO 8601 UTC timestamp
          example: "2026-02-09T10:00:00Z"

    # ==================== Match Schemas ====================
...
    GeoJsonPoint:
      type: object
      required:
        - geometryType
        - coordinates
      properties:
        geometryType:
          type: string
          enum: [Point]
          description: GeoJSON point geometry type
        coordinates:
          type: array
          minItems: 2
          maxItems: 2
          items:
            type: number
          example: [-73.935242, 40.730610]
          description: "[longitude, latitude] format"

    # ==================== Cloud Upload Schemas ====================
    UploadUrlRequest:
      type: object
      description: request for presigned URL to upload file to cloud storage
      required:
        - fileName
        - contentType
      properties:
        fileName:
          type: string
          example: "vacation_hiking.jpg"
        contentType:
          type: string
          example: "image/jpeg"
    UploadUrlResponse:
      type: object
      required:
        - fileId
        - uploadUrl
        - expirationTime
      properties:
        fileId:
          type: string
          format: uuid
        uploadUrl:
          type: string
          format: uri
          description: cloud storage URL for image binary save request
        expirationTime:
          type: string
          format: date-time
          description: ISO 8601 UTC timestamp
          example: "2026-02-08T12:00:00Z"

    # ==================== Error Schemas ====================
    GlobalError:
      type: object
      required:
        - errorCode
        - errorMessage
        - timestamp
        - endpointURI
      properties:
        errorCode:
          type: string
          example: RESOURCE_NOT_FOUND
        errorMessage:
          type: string
          example: The requested hobby was not found.
        timestamp:
          type: string
          format: date-time
        endpointURI:
          type: string
          example: "/api/hobbies/Hiking"
  responses:
...
    BadRequestError: #400
      description: "The request was invalid or cannot be served."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GlobalError'
    UnauthorizedError: #401
      description: "Authentication is required or has failed."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GlobalError'
    ForbiddenError: #403
      description: "Insufficient permissions to access this resource."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GlobalError'
    NotFoundError: #404
      description: "The requested resource could not be found."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GlobalError'
...
    InternalServerError: #500
      description: "An unexpected error occurred on the server."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GlobalError'