# local-hobbies-backend/src/main/java/io/github/csci499_group8/local_hobbies/backend/dto/user/UserOnboardingRequest.java
package io.github.csci499_group8.local_hobbies.backend.dto.user;

import io.github.csci499_group8.local_hobbies.backend.dto.availability.AvailabilityOnboardingRequests;
import io.github.csci499_group8.local_hobbies.backend.dto.common.GeoJsonPoint;
import io.github.csci499_group8.local_hobbies.backend.dto.hobby.HobbyCreationRequest;
import jakarta.validation.Valid;

import java.time.LocalDate;
import java.util.List;

/**
 * Request to complete onboarding; contains only fields that are being updated
 */
public record UserOnboardingRequest(
    String name,
    LocalDate birthDate,
    @Valid GeoJsonPoint location,
    String publicContactInfo,
    GenderMatched genderMatched,
    @Valid List<HobbyCreationRequest> hobbies,
    @Valid AvailabilityOnboardingRequests availabilities
) {}